pyproject.toml
[tool.poetry]
name = "asttrs"
version = "0.1.0"
description = "A attrs-style wrapper for python ast"
authors = ["ryanchao2012 <ryanchao2012@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^21.2.0"
cattrs = "^1.8.0"
astor = "^0.8.1"
black = "^21.9b0"
isort = "^5.9.3"

[tool.poetry.dev-dependencies]
ipython = "^7.28.0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
